requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow-cpu